separate module for specific features definition
