+++ src/biome/text/configuration.py
@@ -4,105 +4,12 @@ from typing import Any, Dict, List, Optional, Type, Union
 from allennlp.common import FromParams, Params
 from allennlp.data import TokenIndexer, Vocabulary
 from allennlp.modules import TextFieldEmbedder
-from biome.text.modules.specs import Seq2VecEncoderSpec
 
+from .features import CharFeatures, WordFeatures
 from .featurizer import InputFeaturizer
 from .modules.encoders import Encoder
 from .modules.heads import TaskHeadSpec
 from .tokenizer import Tokenizer
-
-
-class WordFeatures:
-    """Feature configuration at word level"""
-
-    namespace = "word"
-
-    def __init__(
-        self,
-        embedding_dim: int,
-        lowercase_tokens: bool = False,
-        trainable: bool = True,
-        weights_file: Optional[str] = None,
-        **extra_params
-    ):
-        self.embedding_dim = embedding_dim
-        self.lowercase_tokens = lowercase_tokens
-        self.trainable = trainable
-        self.weights_file = weights_file
-        self.extra_params = extra_params
-
-    @property
-    def config(self):
-        config = {
-            "indexer": {
-                "type": "single_id",
-                "lowercase_tokens": self.lowercase_tokens,
-                "namespace": self.namespace,
-            },
-            "embedder": {
-                "embedding_dim": self.embedding_dim,
-                "vocab_namespace": self.namespace,
-                "trainable": self.trainable,
-                **({"pretrained_file": self.weights_file} if self.weights_file else {}),
-            },
-        }
-
-        for k in self.extra_params:
-            config[k] = {**self.extra_params[k], **config.get(k)}
-
-        return config
-
-    def to_json(self):
-        data = vars(self)
-        data.update(data.pop("extra_params"))
-
-        return data
-
-
-class CharFeatures:
-    """Feature configuration at character level"""
-
-    namespace = "char"
-
-    def __init__(
-        self,
-        embedding_dim: int,
-        encoder: Dict[str, Any],
-        dropout: int = 0.0,
-        **extra_params
-    ):
-        self.embedding_dim = embedding_dim
-        self.encoder = encoder
-        self.dropout = dropout
-        self.extra_params = extra_params
-
-    @property
-    def config(self):
-        config = {
-            "indexer": {"type": "characters", "namespace": self.namespace},
-            "embedder": {
-                "type": "character_encoding",
-                "embedding": {
-                    "embedding_dim": self.embedding_dim,
-                    "vocab_namespace": self.namespace,
-                },
-                "encoder": Seq2VecEncoderSpec(**self.encoder)
-                .input_dim(self.embedding_dim)
-                .config,
-                "dropout": self.dropout,
-            },
-        }
-
-        for k, v in self.extra_params.items():
-            config[k] = {**self.extra_params[k], **config.get(k)}
-
-        return config
-
-    def to_json(self):
-        data = vars(self)
-        data.update(data.pop("extra_params"))
-
-        return data
 
 
 class FeaturesConfiguration(FromParams):
@@ -126,14 +33,12 @@ class FeaturesConfiguration(FromParams):
     
     Parameters
     ----------
-    word : `WordFeatures`
-    char: `CharFeatures`
+    word : `biome.text.features.WordFeatures`
+    char: `biome.text.features.CharFeatures`
     extra_params
     """
 
     __DEFAULT_CONFIG = WordFeatures(embedding_dim=50)
-    __INDEXER_KEYNAME = "indexer"
-    __EMBEDDER_KEYNAME = "embedder"
 
     WORDS = WordFeatures.namespace
     CHARS = CharFeatures.namespace
@@ -178,7 +83,7 @@ class FeaturesConfiguration(FromParams):
         return TextFieldEmbedder.from_params(
             Params(
                 {
-                    feature: config[self.__EMBEDDER_KEYNAME]
+                    feature: config["embedder"]
                     for feature, config in configuration.items()
                 }
             ),
@@ -207,7 +112,7 @@ class FeaturesConfiguration(FromParams):
         configuration = self._make_allennlp_config()
 
         indexer = {
-            feature: TokenIndexer.from_params(Params(config[self.__INDEXER_KEYNAME]))
+            feature: TokenIndexer.from_params(Params(config["indexer"]))
             for feature, config in configuration.items()
         }
         return InputFeaturizer(tokenizer, indexer=indexer)
@@ -342,7 +247,6 @@ class TrainerConfiguration:
         in_memory_batches: int = 2,
         data_bucketing: bool = True,
     ):
-
         self.optimizer = optimizer
         self.validation_metric = validation_metric
         self.patience = patience
